BOOK-139 movendo a tela home para o path correto e corrigindo erros no service
